Adding a pending deletes list
The MVCC system is used to allow 'drop' and 'close' to happen in a
non-blocking way. One side affect of this is that if a client issues
a create/drop/create sequence, the second create can arrive before
the drop has taken place. To make this worse, it's possible that the
create/drop delta updates have been incorporated into the primary set
map. This means it LOOKS like the set does not exist (not in the map or
in the delta set), however the vacuum thread may be waiting on a version
barrier, or applying the delta updates and the set still exists (on
disk). In this case, we still want create to report a delete is in
progress. To support this, we create a lightweight list of pending
deletes that is created before we swap trees, and is cleared after the
vacuum thread does the deletes. This causes a minor slow down in create
since it needs to scan this list, but there should be little contention
for the list.